rename files to not confuse git
