+++ verification/sparse_copy/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_metall_executable(verify_sparse_copy_manual verify_sparse_copy_manual.cpp)
 add_metall_executable(verify_sparse_copy verify_sparse_copy.cpp)
+add_metall_executable(verify_sparse_copy_syscalls verify_sparse_copy_syscalls.cpp)
